+++ src/utils/index.ts
@@ -1,13 +1,6 @@
-export const reflow = () => document.body.clientWidth;
-
-export const preventSmoothScrollOnTabbing = () => {
-  document.addEventListener('keydown', (e) => {
-    if (e.key !== 'Tab') return;
-
-    document.documentElement.style.scrollBehavior = '';
-
-    setTimeout(() => {
-      document.documentElement.style.scrollBehavior = 'smooth';
-    });
-  });
-};
+export * from './navLogoTransitions';
+export * from './parseKeyword';
+export * from './preventSmoothScrollOnTabbing';
+export * from './reflow';
+export * from './rememberSearch';
+export * from './routeReadyState';
