+++ src/utils/preventSmoothScrollOnTabbing.ts
@@ -1,0 +1,11 @@
+export const preventSmoothScrollOnTabbing = () => {
+  document.addEventListener('keydown', (e) => {
+    if (e.key !== 'Tab') return;
+
+    document.documentElement.style.scrollBehavior = '';
+
+    setTimeout(() => {
+      document.documentElement.style.scrollBehavior = 'smooth';
+    });
+  });
+};
