+++ src/utils/reflow.ts
@@ -1,0 +1,1 @@
+export const reflow = () => document.body.clientWidth;
